Create static orr-rules
Legal counsel wanted a static ORR Rules page to appear with the other Legal Pages (Privacy and TOU) in the footer. If there is a better way to do this with the existing show.md please let me know. This version does not need the "I accept" at the bottom since it is just to provide users a link to this documentation.