Fixes CORS issue with video tracks
Changes value of `crossorigin` property from "anonymous" to "with-credentials",
which allows the caption file to be served from S3, and also sends the cookie
containing the session ID so that the MediaController can successfully retrieve
the user's session, which has the correct value for indicating the user has
agreed to the TOS, which allows the video to play.

Fixes #1033.