Adds new ability for direct access to media urls
The same permission check was being used to determine whether to display the
media player in the browser AND to see whether media URLs can be accessed
directly. This was leading to some convoluted permutations in the CanCan Ability
class. It's still pretty convoluted actually, but refactoring started to get
into the weeds, so I tried keeping it minimal.

Also, removes unused argument `name` from Cleaner#clean.