Specify ruby version in Gemfile
By putting it in the Gemfile, if formalizes the ruby version as an actual
dependency, instead of relying on RVM to read .ruby-version. NOTE: RVM may
not be present in prodution environments.

By moving ruby version to Gemfile, .ruby-version becomes redundant, so that is
removed, however that means we have to explicitly specify which Ruby versions to
use in Travis CI on .travis.yml, which I'm ok with.