Gets captions working
Involves a non-trivial refactor of how captions work. Logic from Transcripter
was refactored into 2 different classes: CaptionFile, which is the model that
retrieves raw SRT from S3, and provides accessors to different formats of the
captions file. The CaptionConverter class does the actual transformations from
srt format to webvtt and html.

Then the TranscriptController responds to file extensions for .vtt, .srt, and
.html. For captions in the <video> element, we now simply refer to a URL such
as "/transcripts/cpb-aacip_123-4567.vtt" and it works.

This commit also introduces WebMock which is used to mock request/responses to
S3.

Closes #1054.