Adds video-js back in as the media player.
We were using video-js before, but removed it in favor of using native HTML5,
for the sake of fewer dependencies. However, we lost poster imgs for Audio
playback, and we had inconsistencies with closed-caption button in FF. So we
decided to use video-js again.

Note that we're pulling video-js scripts and styles from their recommended CDN
rather than serving it from our application. I did this because we don't have
a CDN set up for serving assets, and including video-js scripts and styles in
our precompiled assests would bloat them quite a bit. I'm guessing (without
any benchmarks) that serving from their CDN would be faster. This does, however,
make us reliant on the VJS links working for as long as we need them. If it they
turn out to be unreliable, we can always pull a copy down and serve it
ourselves.

Closes #1064. Closes #1043.