Use HLS streaming rather than progressive download
Streaming from Sony Ci makes it a bit harder to capture media from our site, so we are
switching to that.

To do this, we're switching to use the new Sony Ci API gem which provides a cleaner interface to the
underlying Sony Ci API, including an method for fetching HLS streaming URLs, and will allow for more
interaction with Sony Ci in the future.

Also removes a one-off script and spec dependent on the old Sony Ci API gem, but we don't need it anymore.

Closes #2346.