Use HLS streaming for video playback
Streaming from Sony Ci makes it a bit harder to capture media from our site, so we are
switching to using HLS streaming URLs rather than progressive download URLs.

To do this, we're switching to use the new Sony Ci API gem which provides a cleaner interface to the
underlying Sony Ci API, including an method for fetching HLS streaming URLs, and will allow for more
interaction with Sony Ci in the future.

NOTE: Sony Ci does not provide a streaming URL for audio files, so for now those remain as downloadable
URLs.

Also,
* removes a one-off script and spec dependent on the old Sony Ci API gem, but we don't need it anymore.
* guards against empty transcript file content when creating transcipt snippets in catalog#index
* guards against empty JSON param passed to the transcript view helper method

Closes #2356.

Co-authored-by: Andrew Myers <andrew_myers@wgbh.org>