Fixes error when transcript file is missing (cont) (#2412)
Loads the Rails environment in download_clean_ingest.rb instead of stubbing the Rails module.

The use of ExternalFile within the ingest context was brittle because for some
reason thed download_clean_ingest.rb is explicitly avoiding loading the Rails
environment, and instead methods on the Rails module, such as .cache, and
simply calling Rails.cache.delete in the ExternalFile class caused the ingest
script to fail.

Not sure why the Rails environment was deliberately avoided during the ingest script, other
than a way to save on memory. But it's weird, and leads to brittle code, as this case
highlights.