Fixes regex to match beginnig of IP address
The BadBotConfigLine class spits out sample apache configuration based to block certain IP
address blocks based on their frequency in the apache log. The regex in the script was
erroneously matching octet pairs to any part of the clients IP to flag it as a bad bot.
This PR fixes the regex to only match the beginning

Fixes #2899